object the end

+++ taskmanager/main/views.py
@@ -187,6 +187,10 @@ def profile_bb_delete(request, pk):
         messages.add_message(request, messages.SUCCESS,
                              'Чужое!!!! , трогать нельзя')
         return redirect('profile')
+    if not hasattr(Bb.STATUS_CHOISES, 'new'):
+        messages.add_message(request, messages.SUCCESS,
+                             'Вы не можете редактировать это объявление ведь его статус изменен')
+        return redirect('profile')
     if request.method == 'POST':
         bb.delete()
         messages.add_message(request, messages.SUCCESS,
